+++ src/tag/tag.tsx
@@ -77,7 +77,6 @@ export default class Tag extends PureComponent<TagProps> {
 
   componentDidUpdate(prevProps: TagProps) {
     if (this.props.focused !== prevProps.focused) {
-      // eslint-disable-next-line react/no-did-update-set-state
       this.setState({focused: this.props.focused});
     }
     if (this.state.focused) {